refactored components directory so is more understandable now
